backend/IA/requirements.txt
fastapi
uvicorn
pillow
keras
tensorflow
flask
